Fix finding input test files
Under the main block in parse.py, if the script was run anywhere except from the input directory, the input files could not be read. So, simplest method is just to use the current working directory as the base of the path.

Con: the script must then be run from the base directory in order to correctly find the files. But running it from anywhere else is not important (right now; famous last words).